Split out some firecracker types
