+++ pkg/runtimes/firecracker/snapshotter.go
@@ -11,7 +11,6 @@ import (
 	"path/filepath"
 	"slices"
 	"strings"
-	"time"
 
 	iutils "github.com/loopholelabs/drafter/internal/utils"
 
@@ -55,56 +54,6 @@ var (
 	ErrCouldNotCloseAcceptingAgent           = errors.New("could not close accepting agent")
 	ErrCouldNotCreateSnapshot                = errors.New("could not create snapshot")
 )
-
-type PackageConfiguration struct {
-	AgentVSockPort uint32 `json:"agentVSockPort"`
-}
-
-type AgentConfiguration struct {
-	AgentVSockPort uint32
-	ResumeTimeout  time.Duration
-}
-
-type LivenessConfiguration struct {
-	LivenessVSockPort uint32
-	ResumeTimeout     time.Duration
-}
-
-type SnapshotDevice struct {
-	Name   string `json:"name"`
-	Input  string `json:"input"`
-	Output string `json:"output"`
-}
-
-type HypervisorConfiguration struct {
-	FirecrackerBin string
-	JailerBin      string
-
-	ChrootBaseDir string
-
-	UID int
-	GID int
-
-	NetNS         string
-	NumaNode      int
-	CgroupVersion int
-
-	EnableOutput bool
-	EnableInput  bool
-}
-
-type NetworkConfiguration struct {
-	Interface string
-	MAC       string
-}
-
-type VMConfiguration struct {
-	CPUCount    int
-	MemorySize  int
-	CPUTemplate string
-
-	BootArgs string
-}
 
 func CreateSnapshot(ctx context.Context, devices []SnapshotDevice,
 	vmConfiguration VMConfiguration, livenessConfiguration LivenessConfiguration,
