+++ pkg/runtimes/firecracker/types.go
@@ -1,0 +1,53 @@
+package firecracker
+
+import "time"
+
+type PackageConfiguration struct {
+	AgentVSockPort uint32 `json:"agentVSockPort"`
+}
+
+type AgentConfiguration struct {
+	AgentVSockPort uint32
+	ResumeTimeout  time.Duration
+}
+
+type LivenessConfiguration struct {
+	LivenessVSockPort uint32
+	ResumeTimeout     time.Duration
+}
+
+type SnapshotDevice struct {
+	Name   string `json:"name"`
+	Input  string `json:"input"`
+	Output string `json:"output"`
+}
+
+type HypervisorConfiguration struct {
+	FirecrackerBin string
+	JailerBin      string
+
+	ChrootBaseDir string
+
+	UID int
+	GID int
+
+	NetNS         string
+	NumaNode      int
+	CgroupVersion int
+
+	EnableOutput bool
+	EnableInput  bool
+}
+
+type NetworkConfiguration struct {
+	Interface string
+	MAC       string
+}
+
+type VMConfiguration struct {
+	CPUCount    int
+	MemorySize  int
+	CPUTemplate string
+
+	BootArgs string
+}
